Refactors CustomDropDownView
